reset the cursor
