+++ test/development/workload/PageIterator.ts
@@ -8,10 +8,6 @@
 import { Event } from "@microsoft/microsoft-graph-types";
 import { assert } from "chai";
 
-import { Client, ClientOptions } from "../../../src";
-import { ChaosHandler } from "../../../src/middleware/ChaosHandler";
-import { ChaosHandlerOptions } from "../../../src/middleware/options/ChaosHandlerOptions";
-import { ChaosStrategy } from "../../../src/middleware/options/ChaosStrategy";
 import { GraphRequestOptions, PageIterator, PageIteratorCallback } from "../../../src/tasks/PageIterator";
 import { getClient } from "../test-helper";
 const client = getClient();
@@ -91,33 +87,4 @@ describe("PageIterator", () => {
 			assert.isTrue(pageIterator.isComplete());
 		}
 	}).timeout(30 * 1000);
-
-	// TODO - Temporariliy commenting this test.
-	it("setting middleware with pageIterator", async () => {
-		const middleware = new ChaosHandler();
-		const getPageCollection = () => {
-			return {
-				value: [],
-				"@odata.nextLink": "nextURL",
-				additionalContent: "additional content",
-			};
-		};
-		const clientOptions: ClientOptions = {
-			middleware,
-		};
-		const responseBody = { value: [{ event1: "value1" }, { event2: "value2" }] };
-		let counter = 1;
-		const callback: PageIteratorCallback = (data) => {
-			assert.equal(data["event" + counter], "value" + counter);
-			counter++;
-			return true;
-		};
-
-		const middlewareOptions = [new ChaosHandlerOptions(ChaosStrategy.MANUAL, "middleware options for pageIterator", 200, 0, JSON.stringify(responseBody), new Headers({ "Content-Type": "application/json", "content-length": "100" }))];
-		const requestOptions = { middlewareOptions };
-
-		const client = Client.initWithMiddleware(clientOptions);
-		const pageIterator = new PageIterator(client, getPageCollection(), callback, requestOptions);
-		await pageIterator.iterate();
-	});
 });
